feat: refactor plugin file name (#539)
